Make requests.date_received nullable
The idea of this column is that it can be used to record the date
a request was received, which may be earlier than the date it was
entered into the system.

We make it nullable, to distinguish cases where a received date was
entered explicitly from cases where it was not entered and is assumed
to be just the date the database record was created.