Authentication
So the idea is that you can create a new staff member at

/staff_members/new

and this is protected only by basic auth, currently using a
hard-coded username and password.

A staff member can log in at

/sessions/new

and log out at

/sessions/logout

which then gives access to the main app. If you try to access the
app without being logged in, you are redirected to the login page.