State descriptions
Improve the names of the exempt states so they include a clue
to the nature of the different numbered exemptions.

Show the state description when creating or editing a request,
both as tooltip text on the state dropdown and also to the right
of the dropdown once a state has been selected.